modify scripts to streamline lifter testing

+++ get_opcode.sh
@@ -5,8 +5,14 @@ initial_dir=$(pwd)
 tmp_bin="/tmp/getopcode_bin_"$USER
 
 cd "${script_dir}"
-echo "$@" | aarch64-linux-gnu-as - -o "${tmp_bin}" &&
-objdump "${tmp_bin}" -D | python3 get_opcode.py
 
+echo "$@" | aarch64-linux-gnu-as -march=armv8.5-a - -o "${tmp_bin}"
+return_code=$?
+if [ "$return_code" -ne 0 ]; then
+    cd "${initial_dir}"
+    exit "$return_code"
+fi
+
+objdump "${tmp_bin}" -D | python3 get_opcode.py &&
 rm "${tmp_bin}"
 cd "${initial_dir}"
